+++ src/condor_dagman/debug.cpp
@@ -281,7 +281,7 @@ time_to_str( time_t timestamp, std::string &tstr )
 	// function, but this is a quick hack for LIGO. I'll come back to it
 	// and do it better later when I have time.
 	if ( UseTimestamps ) {
-		formatstr( tstr, "(%ld) ", timestamp );
+		formatstr( tstr, "(%lld) ", (long long)timestamp );
 	} else {
 		struct tm *tm = nullptr;
 		tm = localtime( &timestamp );
