+++ src/condor_utils/directory.cpp
@@ -1134,7 +1134,7 @@ create_temp_file(bool create_as_subdirectory) {
 	int retry_count = 10;
 
 	do {
-		snprintf (filename, 500, "%s/tmp.%d.%ld.%d", temp_dir, mypid, timestamp++, counter++);
+		snprintf (filename, 500, "%s/tmp.%d.%lld.%d", temp_dir, mypid, (long long)timestamp++, counter++);
 		filename[500-1] = 0;
 	} while ((--retry_count > 0) && 
 			 ( (!create_as_subdirectory && (fd=safe_open_wrapper_follow(filename, O_EXCL | O_CREAT, S_IREAD | S_IWRITE)) == -1) ||
